+++ tests/testthat/test_api_ollama.R
@@ -114,8 +114,22 @@ test_that("ollama_embedding returns expected response", {
                 "The limits of my language mean the limits of my world") |>
       ollama_embedding()
     
-    expect_true(is.numeric(result))
-    expect_equal(dim(result),c(384,4))
+    
+    # Test that the result is a tibble
+    expect_s3_class(result, "tbl_df")
+    
+    # Test that the tibble has two columns: input and embeddings
+    expect_named(result, c("input", "embeddings"))
+    
+    # Test that the input column contains the original input texts
+    expect_equal(result$input, c("It is not that I am mad, it is only that my head is different from yours",
+                                 "A man can do as he wills, but not will as he wills",
+                                 "Whereof one cannot speak, thereof one must be silent",
+                                 "The limits of my language mean the limits of my world"))
+    
+    purrr::walk(result$embeddings, function(embedding) {
+      expect_equal(length(embedding), 384)
+    })
     
   },simplify = FALSE)
 })